feat: unit preference buttons
Also fixes overflowing text in the table and in the current weather
display.
The photo for rainy weather has been darkened for better contrast and
the photos for clear and cloudy nights have been swapped in favour of
landscape sized images.
Small changes to the forecast data table, including separators and data
for the current hour.